requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
dj-database-url==2.3.0
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.5.0
git-filter-repo==2.47.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
multidict==6.1.0
packaging==24.2
propcache==0.2.1
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.0
pyOpenSSL==24.3.0
python-dotenv==1.0.1
requests==2.32.3
service-identity==24.2.0
setuptools==75.6.0
sqlparse==0.5.2
twilio==9.3.7
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.8.2
yarl==1.18.3
zope.interface==7.2
